Revert "Baan Modal & Refactor Modal"
