--- src/components/rpkm/Modal/ModalButton.tsx
@@ -1,41 +1,0 @@
-import { cn } from '@/lib/utils';
-
-interface ModalButtonProps {
-  callBackFunction: (params: unknown) => void;
-  borderClassName: string;
-  backgroundClassName: string;
-  children: React.ReactNode;
-}
-
-/**
- * ModalButton component
- * @param callBackFunction - function
- * @param borderClassName - string
- * @param backgroundClassName - string
- * @param children - ReactNode
- * @returns Styled button component
- */
-const ModalButton: React.FC<ModalButtonProps> = ({
-  callBackFunction,
-  borderClassName,
-  backgroundClassName,
-  children,
-}) => {
-  return (
-    <button
-      onClick={callBackFunction}
-      className={cn('p-1 inv-rad inv-rad-2 w-2/5', borderClassName)}
-    >
-      <div
-        className={cn(
-          'py-[0.3rem] px-2 inv-rad inv-rad-2',
-          backgroundClassName
-        )}
-      >
-        {children}
-      </div>
-    </button>
-  );
-};
-
-export default ModalButton;
