--- src/components/rpkm/Modal/ModalStyle.ts
@@ -1,22 +1,0 @@
-const modalStyles = {
-  red: {
-    background: 'bg-[#C94B4B]',
-    button: {
-      'accept-background': 'bg-[#183F86] text-[#EAE3C3]',
-      'accept-border': 'bg-[#F5F5F5]',
-      'cancel-background': 'bg-[#F5F5F5] text-[#414643]',
-      'cancel-border': 'bg-[#414643]',
-    },
-  },
-  blue: {
-    background: 'bg-[#183F86]',
-    button: {
-      'accept-background': 'bg-[#C94B4B] text-[#EAE3C3]',
-      'accept-border': 'bg-[#EAE3C3]',
-      'cancel-background': 'bg-[#313131] text-[#F5F5F5]',
-      'cancel-border': 'bg-[#F5F5F5]',
-    },
-  },
-};
-
-export default modalStyles;
